Qualify file directorys as either Plugin files or Main Application file
Documentation just says add the code [line] in config/bootstrap.php without specifying whether this is app_name/config/bootstrap.php or .../cakedc/users/config/bootstrap.php. This also goes for users. 

As a first time user of this plugin I created a users.php file in app_name/config/... and as I'm writing this am not sure which directory to put my code in. 

This also goes for the rest of the documentation. 

Thanks.